change version

+++ backend/src/app/modules/newsLetter/newsLetter.controller.ts
@@ -2,11 +2,16 @@ import { Request, Response } from "express";
 import catchAsync from "../../shared/catchAsync";
 import { newsletterService } from "./newsLetter.service";
 import sendResponse from "../../shared/sendResponse";
+import AppError from "../../errors/appError";
+import { HttpStatus } from "http-status";
 
 
 
 const subscribeController = catchAsync(async (req: Request, res: Response) => {
   const { email } = req.body;
+    if (!email || typeof email !== "string") {
+    throw new AppError(402, "Valid email is required");
+  }
   const subscriber = await newsletterService.subscribeNewsletter(email);
   sendResponse(res, {
     statusCode: 201,
@@ -31,6 +36,9 @@ const listSubscribersController = catchAsync(async (req: Request & { user?: any 
 
 const deleteSubscriberController = catchAsync(async (req: Request, res: Response) => {
   const { id } = req.params;
+  if (!id || typeof id !== "string") {
+    throw new AppError(402, "Subscriber ID is required");
+  }
   const subscriber = await newsletterService.removeSubscriber(id);
   sendResponse(res, {
     statusCode: 200,
